Feat: Move code autocomplete logic to Autocomplete.Code module
